feature(select-component): scroll down action
