+++ challenge-catar Watch App/Views/ContentView.swift
@@ -33,17 +33,7 @@ class ScreenObserver: ObservableObject {
 
 
 
-class ScrollDown: ScrollableDirection {
-    
-    func execute(with screens: [ScreenAvaiable], andCurrent scrolling: Int) -> ([ScreenAvaiable], Int) {
-        var newScreens = screens
-        var newScrolling = scrolling
-        if newScrolling != screens.count - 1 { newScreens[scrolling].height = 82 }
-        newScrolling += 1
-        if newScrolling == screens.count { newScrolling -= 1 }
-        return (newScreens, newScrolling)
-    }
-}
+
 
 class ScrollFactory {
